Implementing the basics of routes without error handling and fixing database integration

+++ GpsPoiApp/Services/IPointService.cs
@@ -6,5 +6,5 @@ public interface IPointService
 {
     Point AddPoint(Point point);
     List<Point> GetAllPoints();
-    List<Point> GetPointsByProximity(int x, int y, int maxDistance);
+    List<Point> GetPointsByProximity(int x, int y, double maxDistance);
 }
